% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separar-apellido.R
\name{separar_apellido}
\alias{separar_apellido}
\title{Separa la columna apellido}
\usage{
separar_apellido(data, col = NULL)
}
\arguments{
\item{data}{Un dataframe.}

\item{col}{Opcional, especificacion de columna con tidyselect.}
}
\value{
Un dataframe con solo la columna apellido.
}
\description{
Separa la columna apellido de un dataframe.
}
\examples{
# separar_apellido(clientes)




}
